Summary: Major refactoring and feature additions in LC3b project
This commit includes a significant overhaul of the LC3b project. The ALU.cs file was removed, and new classes `InstructionDecoder` and `ProgramCounter` were added. The `Main` method in Program.cs was heavily modified to include instruction set initialization, execution of loaded instructions, and display of general purpose registers content.

New methods for initializing the instruction set and loading instructions from a file were also added. The parameters of the `Execute` method in several instruction classes were changed to be passed by reference.

In Memory.cs, a new method `UsedStorage` was added and the `Loader` property was moved to a new region. In Registers.cs and Loader.cs, properties were moved to a new region named `Eigenschaften`.

The instruction_file.txt was updated with new instructions. The project names and IDs in LC3b.sln were changed. A new test class `InstructionTests` was added in InstructionTests.cs, and a new project file TestLC3b.csproj was added for the test project.